requirements.txt
fastapi
uvicorn
pydantic
langchain-core
langchain-community
langchain-groq
streamlit
deepgram-sdk
